Struct steering::SteeringAcceleration

pub struct SteeringAcceleration<T: Real> {
    pub linear: Vector3<T>,
    pub angular: T,
}

Represents result of a steering behaviour computation. User can aggregate more than one behaviour result into single acceleration struct.

Fields
linear: Vector3<T>

linear acceleration component

angular: T

angular acceleration component

Methods

impl<T: Real> SteeringAcceleration<T>

pub fn default() -> SteeringAcceleration<T>

pub fn new(
linear_acceleration: Vector3<T>,
angular_acceleration: T
) -> SteeringAcceleration<T>

Creates a steering acceleration struct using given linear and angular components

pub fn is_zero(&self) -> bool

Tests whether both linear and angular acceleration compenents are zero

pub fn set_zero(&mut self) -> &mut Self

Sets both compononents to zero

pub fn add(&mut self, other: SteeringAcceleration<T>) -> &mut Self

pub fn scl(&mut self, scale: T) -> &mut Self

pub fn mul_add(&mut self, other: SteeringAcceleration<T>, scale: T) -> &mut Self

pub fn calculate_square_magnitude(&self) -> T

pub fn calculate_magnitude(&self) -> T